도깨비불 추가 및 회수기능 추가 (커신이 떠날때마다 떨구며 레어도에 따라 주는 정도가 달라짐) 일반 커신 3종 추가 (이제 레어가 레어다워짐 드릅게 안나와ㅡㅡ) 아이템 이미지랑 설명 변경 UI 살짝 변경함 파티클 잣버그 수정 (이거 투명도 다시 안되돌려서 생긴 문제였음) 기타 잣버그들 수정

+++ Assets/scripts/WispInfo.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public class WispInfo : MonoBehaviour
 {
     public Sprite wispImage;
-    public float wispSize;
+    public long wispSize;
 }
